Extracted a 'Contact' class
